uf-10296: Moved cache tests to correct test class.
